refactor(test): share session cookie assertions across integration tests
